requirements.txt
python-telegram-bot==21.3
SQLAlchemy==2.0.31
httpx==0.27.0